+++ src/vim/table-embedded-editor.ts
@@ -1,9 +1,10 @@
 import {
     setTableNavControllerEnabled,
+    setTableNavWhichKeyConfig,
     tableNavControllerField,
 } from './table-nav-controller';
 import { setTableEmbeddedMode } from './table-render-widget';
 
-export { setTableEmbeddedMode };
+export { setTableEmbeddedMode, setTableNavWhichKeyConfig };
 export const setEmbeddedModeEnabled = setTableNavControllerEnabled;
 export const tableEmbeddedField = tableNavControllerField;